`u2s` and `s2u` virtual instructions
convert an unsigned number to a signed number and a signed number to an unsigned number. because signed range is smaller than the unsigned range, high bits of the register may be lost. when converting from a signed number to an unsigned number, the sign bit will be discarded.
casts between unsigned and signed numbers aren't necessary.